Isolate real hook worker coverage
